fix integer cast possible undefined behavior
Maybe fdb_iteration should panic or handle the error itself.
`iteration` is checked in foundationdb and returns client_invalid_operation if it is <= 0
There is no information in the API spec about it.
For example if we pass a key.len() < 0 as we could before this commit, it is undefined behavior, because foundationdb is not checking it.

+++ foundationdb/src/fdb_api.rs
@@ -20,7 +20,7 @@ use foundationdb_sys as fdb_sys;
 
 /// Returns the max api version of the underlying Fdb C API Client
 pub fn get_max_api_version() -> i32 {
-    unsafe { fdb_sys::fdb_get_max_api_version() as i32 }
+    unsafe { fdb_sys::fdb_get_max_api_version() }
 }
 
 static VERSION_SELECTED: AtomicBool = AtomicBool::new(false);
